chinese_compostion_desciption/module/description_recognition/requirement.txt
torch==1.2.0
pytorch_transformers==1.2.0
jieba==0.39
numpy==1.17.2
scikit-learn==0.20.2